Relocate HttpClientService class
